Move files to lib
