--- vscode/packages/npm-package/index.ts
@@ -1,13 +1,0 @@
-export * from './config';
-export * from './lib/generateClient';
-export * from './lib/generatePlugin';
-export * from './lib/getKiotaTree';
-export * from './lib/getKiotaVersion';
-export * from './lib/getManifestDetails';
-export * from './lib/languageInformation';
-export * from './lib/migrateFromLockFile';
-export * from './lib/removeItem';
-export * from './lib/searchDescription';
-export * from './lib/updateClients';
-export * from './types';
-export * from './utils';
