+++ vscode/packages/npm-package/lib/config.ts
