+++ vscode/packages/npm-package/lib/connect.ts
