+++ vscode/packages/npm-package/lib/index.ts
@@ -1,0 +1,13 @@
+export * from './config.js';
+export * from './generateClient.js';
+export * from './generatePlugin.js';
+export * from './getKiotaTree.js';
+export * from './getKiotaVersion.js';
+export * from './getManifestDetails.js';
+export * from './languageInformation.js';
+export * from './migrateFromLockFile.js';
+export * from './removeItem.js';
+export * from './searchDescription.js';
+export * from './updateClients.js';
+export * from './types.js';
+export * from './utils.js';
