+++ vscode/packages/npm-package/lib/types.ts
