+++ vscode/packages/npm-package/lib/utils.ts
@@ -1,4 +1,4 @@
-import { KiotaLogEntry, MaturityLevel, DependencyType, LogLevel } from "./types";
+import { KiotaLogEntry, MaturityLevel, DependencyType, LogLevel } from "../types.js";
 
 export function getLogEntriesForLevel(logEntries: KiotaLogEntry[], ...levels: LogLevel[]): KiotaLogEntry[] {
   return logEntries.filter((entry) => levels.indexOf(entry.level) !== -1);
